+++ packages/core/src/model/TransactionOptions.ts
@@ -1,0 +1,6 @@
+import { Signer } from 'ethers'
+
+export interface TransactionOptions {
+  signer?: Signer
+  transactionName?: string
+}