+++ packages/core/src/model/config/default.ts
@@ -3,7 +3,17 @@ import { FullConfig } from './Config'
 
 export const DEFAULT_CONFIG: FullConfig = {
   pollingInterval: 15000,
-  supportedChains: [ChainId.Mainnet, ChainId.Goerli, ChainId.Kovan, ChainId.Rinkeby, ChainId.Ropsten, ChainId.xDai, ChainId.Localhost, ChainId.Hardhat],
+  supportedChains: [
+    ChainId.Mainnet,
+    ChainId.Goerli,
+    ChainId.Kovan,
+    ChainId.Rinkeby,
+    ChainId.Ropsten,
+    ChainId.BSC,
+    ChainId.xDai,
+    ChainId.Localhost,
+    ChainId.Hardhat
+  ],
   notifications: {
     checkInterval: 500,
     expirationPeriod: 5000,