+++ packages/example/src/pages/Balance.tsx
@@ -2,15 +2,16 @@ import React from 'react'
 import { formatEther } from '@ethersproject/units'
 import { useEtherBalance, useEthers } from '@usedapp/core'
 import { Container, ContentBlock, ContentRow, MainContent, Section, SectionRow } from '../components/base/base'
-import { Button } from '../components/base/Button'
 import { Label } from '../typography/Label'
 import { TextInline } from '../typography/Text'
 import { Title } from '../typography/Title'
 
+import { AccountButton } from '../components/account/AccountButton'
+
 const STAKING_CONTRACT = '0x00000000219ab540356cBB839Cbe05303d7705Fa'
 
 export function Balance() {
-  const { activateBrowserWallet, deactivate, account } = useEthers()
+  const { account } = useEthers()
   const userBalance = useEtherBalance(account)
   const stakingBalance = useEtherBalance(STAKING_CONTRACT)
 
@@ -20,11 +21,7 @@ export function Balance() {
         <Section>
           <SectionRow>
             <Title>Balance</Title>
-            {account ? (
-              <Button onClick={() => deactivate()}>Disconnect</Button>
-            ) : (
-              <Button onClick={() => activateBrowserWallet()}>Connect</Button>
-            )}
+            <AccountButton />
           </SectionRow>
           <ContentBlock>
             {stakingBalance && (